Add separate exports for node and wasm
