+++ src/tree-sitter/index.ts
@@ -1,3 +1,2 @@
 export * from './ExpressionBuilder.js'
-export * from './NodeParserAdapter.js'
 export * from './types.js'
